backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=1.4.0
pydantic>=1.10.0
python-dotenv>=1.0.0
openai>=0.27.0
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 机器学习和文本处理
scikit-learn==1.3.2
jieba==0.42.1
numpy==1.26.4

# 工具库
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
